rename + restructure files more sensibly
