+++ setup.py
@@ -4,6 +4,6 @@ setup(
     name='pancancer_evaluation',
     author='Jake Crawford',
     version='0.0.1',
-    description='Utility functions for TCGA pan-cancer evaluation',
-    packages=['pancancer_utilities']
+    description='TCGA pan-cancer evaluation of driver mutation prediction',
+    packages=['pancancer_evaluation']
 )
